Mensaje de "estas seguro?", y de confirmacion de que se ha eliminado

+++ src/main/java/controller/ControllerImplementation.java
@@ -438,7 +438,11 @@ public class ControllerImplementation implements IController, ActionListener {
     public void delete(Person p) {
         try {
             if (dao.read(p) != null) {
-                dao.delete(p);
+                int opc = JOptionPane.showConfirmDialog(delete, "Are you sure you want to delete this person?");
+                if(opc == javax.swing.JOptionPane.YES_NO_OPTION){
+                    dao.delete(p);
+                JOptionPane.showMessageDialog(delete, "Person deleted successfully!" , delete.getTitle(), JOptionPane.INFORMATION_MESSAGE);
+                }
             } else {
                 throw new PersonException(p.getNif() + " is not registered and can not "
                         + "be DELETED");
